WIP: bully protocol.
Leader election implemented.

Missing: start up support, and support for transferring leaders.

For example, if we're editing a page and then a new client connects which
would be elected the new leader, we really need to initially ask the
*current* leader what the page content is, and *then* add ourselves to
the peer pool (allowing ourselves to become the new leader).  This should
work even if the current leader is reloaded; ie, when we ask the current
leader about the content, the other peers may need to have an election
*without us* in order to figure out who should reply with the initialization
data.