Better handle race conditions during form initialization.
Add a server-generated timestamp to the `init-connection` message, and use
this to identify the "oldest session" when arbitrating between multiple
received `form-init` messages.  Also defer updates between the time
that `form-init` is sent and when it is received echoed back from the
hub, in order to ensure that the newly-inited peer does not start up
with stale state.